rename: Create Header directory and move components related to Header
