+++ components/Header/CalendarDropDown.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
 
-import { useDropDown } from '../lib/hooks/useDropDown';
-
-import DropDown from './DropDown';
+import { useDropDown } from '../../lib/hooks/useDropDown';
+import DropDown from '../DropDown';
 
 export default function CalendarDropDown({ title }: { title: string }) {
     const { dropDownRef, isOpen, openDropDown, closeDropDown } = useDropDown();
