+++ components/Header/CalendarModal.tsx
@@ -1,8 +1,9 @@
 import React from 'react';
 
+import { MODAL_NAMES } from '../ModalContainer';
+import ModalFrame from '../ModalFrame';
+
 import styles from './CalendarModal.module.scss';
-import { MODAL_NAMES } from './ModalContainer';
-import ModalFrame from './ModalFrame';
 
 export default function CalendarModal() {
     return (
