+++ components/Header/HelpDropDown.tsx
@@ -1,10 +1,9 @@
 import Image from 'next/image';
 import React from 'react';
 
-import { useDropDown } from '../lib/hooks/useDropDown';
-import help_icon from '../public/images/help_icon.svg';
-
-import DropDown from './DropDown';
+import { useDropDown } from '../../lib/hooks/useDropDown';
+import help_icon from '../../public/images/help_icon.svg';
+import DropDown from '../DropDown';
 
 export default function HelpDropDown() {
     const { dropDownRef, isOpen, openDropDown, closeDropDown } = useDropDown();
