+++ components/Header/SearchCategoryDropDown.tsx
@@ -1,8 +1,7 @@
 import React, { useState } from 'react';
 
-import { useDropDown } from '../lib/hooks/useDropDown';
-
-import DropDown from './DropDown';
+import { useDropDown } from '../../lib/hooks/useDropDown';
+import DropDown from '../DropDown';
 
 export default function SearchCategoryDropDown() {
     const { dropDownRef, isOpen, openDropDown, closeDropDown } = useDropDown();
@@ -12,10 +11,6 @@ export default function SearchCategoryDropDown() {
         closeDropDown();
     };
 
-    // dropdown outerClick doesn't work due to e.stopPropagation() of searchModal
-    // change the way of handling backdrop at Modal? (use window.addEventListener)
-    // or change dropdown? (will become similar to Modal)
-    // any other way?
     return (
         <>
             <div ref={dropDownRef} style={{ height: '100%' }}>
