+++ components/Header/SearchDetailsModal.module.scss
